SERIAL: Detect corruption using a regex
This looks at the column headers at startup to dynamically build a regex
to validate incoming lines. For known columns, this uses a specific
regex and for unknown columns, it just uses a generic match-anything
regex.

When a line does not match the expected format, which typically
indicates some corruption on the serial line, the entire line is
dropped and a message is logged.