SECURITY: Store JWT secret in a file
Previously, it was assumed to be passed in an environment variable, but
this no longer fits the execution model we are using. Instead, the
secret is now stored in a file and autogenerated when it is not present.
If reading or writing the file fails, a temporary secret is generated
that allows authenticating during the session, but whose tokens expire
after the connector quits.